mu4e: add new segment that shows unread emails
This is a new segment for `mu4e-alert' that shows unread emails in the
modeline. For now, we enable it where github notifications are enabled
and make it a simple number.

When experimenting with adding an email icon, the modeline seemed a bit
too crowed so there is no icon for this number. We can perhaps add a
flag later to toggle that. Alternatively, we could expose more options
for the `doom-modeline-mu4e-formatter' that could tweak the number and
the icon.