Force redisplay to trigger height calculations
ertain functions, including e.g. `fit-window-to-buffer', base
their size calculations on values that are incorrect if the
mode-line has a height different from that of the `default' face
and certain other calculations have not yet taken place for the
window in question.

Force a redisplay before `fit-window-to-buffer' using an advice,
like we already did for `resize-temp-buffer-window'.

Close #183.

Reference:
https://github.com/tarsius/moody/issues/19
https://github.com/magit/transient/issues/63