Fix Eglot update hook breaking Eglot shutdown
doom-modeline is using various Eglot internals to display its Eglot
segment. One of those internals, eglot--current-server, recently changed
so that it now triggers a cache variable, eglot--cached-current-server,
to be set.

The hook used by doom-modeline is run during shutdown, after the buffer
becomes unmanaged by Eglot, but before all of its internal state has
been updated. The call to eglot--current-server then triggers the cache
variable to be set to an outdated value, which remains after shutdown
completes.

Any further calls to Eglot will end up seeing the incorrectly cached
value, which causes all kinds of breakage.

Note that, with or without this fix, the implementation relies on Eglot
internals and is likely to be fragile.