Fix call of `define-obsolete-variable-alias'
* doom-modeline-core.el (doom-modeline-check-simple-format): Move
`defcustom' after the `define-obsolete-variable-alias' form.  From
the docstring:

If CURRENT-NAME is a defcustom or a defvar (more generally, any
variable where OBSOLETE-NAME may be set, e.g. in an init file,
before the alias is defined), then the
define-obsolete-variable-alias statement should be evaluated
before the defcustom, if user customizations are to be respected.