Introduce analogue clock time icon
With the SVG library (bundled as part of Emacs), we can generate an
analogue clock for the current time to-the-minute, instead of using the
nearest-hour nerdicon.

This relies on a graphical display, and so a new variable is introduced,
`doom-modeline-time-analogue-clock' and the behaviour is gated
on (display-graphic-p).

There are a number of variables that affect the style of the displayed
clock. For now, I've chosen to be conservative and just expose
`doom-modeline-time-clock-minute-resolution' as a custom variable,
leaving the rest as constants. We may well want to revisit this in the
future.

I've personally been running this variant of the time segment for over a
year now, and as a result am pretty confident there are no behaviour or
performance issues with this implementation.

This implementation is originally sourced from
https://github.com/tecosaur/emacs-config/blob/a3c3b73a00f4e034ffeae11f20b00d1eb948a6e6/config.org#time.